♻️ [#59] refactor: 체형 분석 API 요청 수정"
체형 분석 API에서 요청 받는 DTO를 수정하였습니다.
