+++ src/main/java/com/example/mody/domain/bodytype/service/memberbodytype/MemberBodyTypeCommandService.java
@@ -1,6 +1,6 @@
 package com.example.mody.domain.bodytype.service.memberbodytype;
 
-import com.example.mody.domain.bodytype.dto.BodyTypeAnalysisResponse;
+import com.example.mody.domain.bodytype.dto.response.BodyTypeAnalysisResponse;
 import com.example.mody.domain.member.entity.Member;
 
 public interface MemberBodyTypeCommandService {
